pyproject.toml
[project]
name = "mokumoku-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "satodaiki", email = "banana5388@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.1",
    "discord-py>=2.6.4",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "streamlit>=1.53.0",
]

[project.scripts]
serve = "mokumoku_bot:main"

[build-system]
requires = ["uv_build>=0.9.25,<0.10.0"]
build-backend = "uv_build"
